refactor(googledrive): split routes.py into routes/ package
Convert the 942-line routes.py into four focused sub-modules:
deps.py (shared helpers), oauth.py (4 OAuth endpoints),
folders.py (6 folder endpoints), and files.py (8 file + browse
endpoints). The package __init__.py assembles and re-exports router
so all existing import sites remain unchanged.

Update test patch paths from routes.* to routes.deps.*,
routes.oauth.*, routes.folders.*, and routes.files.* to match
where each symbol now lives.
