+++ app/core_plugins/googledrive/routes/folders.py
@@ -1,0 +1,307 @@
+"""Google Drive folder endpoints."""
+
+import logging
+from datetime import datetime, timezone
+
+from fastapi import APIRouter, BackgroundTasks, Depends, HTTPException, Query, status
+from sqlalchemy import func
+from sqlmodel import Session, select
+
+from app.core.db import get_session
+from app.core_plugins.googledrive.client import GoogleDriveClient
+from app.core_plugins.googledrive.routes.deps import (
+    _sync_folder_files,
+    get_drive_credentials,
+    get_valid_access_token,
+    require_user_id,
+)
+from app.core_plugins.googledrive.types import (
+    CreateFolderRequest,
+    DriveFileResponse,
+    DriveFolderResponse,
+    DriveFolderWithFilesResponse,
+    PaginatedResponse,
+    SyncFolderRequest,
+    SyncStatusResponse,
+)
+from app.core_plugins.googledrive.utils import process_folder_rag
+from app.models.drive import DriveFile, DriveFolder
+
+router = APIRouter()
+logger = logging.getLogger(__name__)
+
+
+@router.get("/folders", response_model=PaginatedResponse[DriveFolderResponse])
+def list_folders(
+    skip: int = Query(0, ge=0, description="Number of items to skip"),
+    limit: int = Query(20, ge=1, le=100, description="Number of items to return"),
+    user_id: int = Depends(require_user_id),
+    session: Session = Depends(get_session),
+) -> PaginatedResponse[DriveFolderResponse]:
+    """List synced Google Drive folders for the current user with pagination."""
+    file_count_subq = (
+        select(DriveFile.folder_id, func.count(DriveFile.id).label("file_count"))  # type: ignore[arg-type]
+        .where(DriveFile.is_deleted == False)  # noqa: E712
+        .group_by(DriveFile.folder_id)  # type: ignore[arg-type]
+        .subquery()
+    )
+
+    base_stmt = (
+        select(DriveFolder, file_count_subq.c.file_count)
+        .outerjoin(file_count_subq, DriveFolder.id == file_count_subq.c.folder_id)  # type: ignore[arg-type]
+        .where(
+            DriveFolder.user_id == user_id,
+            DriveFolder.is_deleted == False,  # noqa: E712
+        )
+    )
+
+    count_stmt = select(func.count()).select_from(
+        select(DriveFolder.id)
+        .where(
+            DriveFolder.user_id == user_id,
+            DriveFolder.is_deleted == False,  # noqa: E712
+        )
+        .subquery()
+    )
+    total = session.exec(count_stmt).one()
+    rows = session.exec(base_stmt.offset(skip).limit(limit)).all()
+
+    return PaginatedResponse(
+        items=[
+            DriveFolderResponse(
+                id=folder.id,  # type: ignore[arg-type]
+                drive_folder_id=folder.drive_folder_id,
+                name=folder.drive_folder_name,
+                parent_id=folder.drive_parent_id,
+                file_count=file_count or 0,
+                last_synced_at=folder.last_synced_at,
+                sync_status=folder.sync_status,
+            )
+            for folder, file_count in rows
+        ],
+        total=total,
+        skip=skip,
+        limit=limit,
+    )
+
+
+@router.post("/folders/sync", response_model=DriveFolderResponse)
+async def sync_folder(
+    request: SyncFolderRequest,
+    background_tasks: BackgroundTasks,
+    user_id: int = Depends(require_user_id),
+    session: Session = Depends(get_session),
+) -> DriveFolderResponse:
+    """Sync an existing Google Drive folder."""
+    client_id, client_secret, _ = get_drive_credentials()
+    access_token = await get_valid_access_token(session, user_id, client_id, client_secret)
+
+    existing = session.exec(
+        select(DriveFolder).where(
+            DriveFolder.user_id == user_id,
+            DriveFolder.drive_folder_id == request.drive_folder_id,
+            DriveFolder.is_deleted == False,  # noqa: E712
+        )
+    ).first()
+    if existing:
+        raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail="Folder is already synced")
+
+    async with GoogleDriveClient(access_token) as client:
+        folder_metadata = await client.get_folder(request.drive_folder_id)
+
+    now = datetime.now(timezone.utc)
+    folder = DriveFolder(
+        user_id=user_id,
+        drive_folder_id=folder_metadata["id"],
+        drive_folder_name=folder_metadata["name"],
+        drive_parent_id=folder_metadata.get("parents", [None])[0] if folder_metadata.get("parents") else None,
+        last_synced_at=now,
+        sync_status="syncing",
+    )
+    session.add(folder)
+    session.commit()
+    session.refresh(folder)
+
+    file_count = await _sync_folder_files(session, folder, user_id, access_token)
+    background_tasks.add_task(process_folder_rag, folder.id, user_id, access_token)  # type: ignore[arg-type]
+
+    return DriveFolderResponse(
+        id=folder.id,  # type: ignore[arg-type]
+        drive_folder_id=folder.drive_folder_id,
+        name=folder.drive_folder_name,
+        parent_id=folder.drive_parent_id,
+        file_count=file_count,
+        last_synced_at=folder.last_synced_at,
+        sync_status=folder.sync_status,
+    )
+
+
+@router.post("/folders", response_model=DriveFolderResponse)
+async def create_folder(
+    request: CreateFolderRequest,
+    user_id: int = Depends(require_user_id),
+    session: Session = Depends(get_session),
+) -> DriveFolderResponse:
+    """Create a new folder in Google Drive and sync it."""
+    client_id, client_secret, _ = get_drive_credentials()
+    access_token = await get_valid_access_token(session, user_id, client_id, client_secret)
+
+    async with GoogleDriveClient(access_token) as client:
+        folder_metadata = await client.create_folder(request.name, request.parent_id)
+
+    now = datetime.now(timezone.utc)
+    folder = DriveFolder(
+        user_id=user_id,
+        drive_folder_id=folder_metadata["id"],
+        drive_folder_name=folder_metadata["name"],
+        drive_parent_id=request.parent_id,
+        last_synced_at=now,
+        sync_status="synced",
+    )
+    session.add(folder)
+    session.commit()
+    session.refresh(folder)
+
+    return DriveFolderResponse(
+        id=folder.id,  # type: ignore[arg-type]
+        drive_folder_id=folder.drive_folder_id,
+        name=folder.drive_folder_name,
+        parent_id=folder.drive_parent_id,
+        file_count=0,
+        last_synced_at=folder.last_synced_at,
+        sync_status=folder.sync_status,
+    )
+
+
+@router.get("/folders/{folder_id}", response_model=DriveFolderWithFilesResponse)
+def get_folder(
+    folder_id: int,
+    user_id: int = Depends(require_user_id),
+    session: Session = Depends(get_session),
+) -> DriveFolderWithFilesResponse:
+    """Get a synced folder with its files."""
+    folder = session.exec(
+        select(DriveFolder).where(
+            DriveFolder.id == folder_id,
+            DriveFolder.user_id == user_id,
+            DriveFolder.is_deleted == False,  # noqa: E712
+        )
+    ).first()
+    if not folder:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Folder not found")
+
+    folder_files = session.exec(
+        select(DriveFile).where(
+            DriveFile.folder_id == folder.id,
+            DriveFile.is_deleted == False,  # noqa: E712
+        )
+    ).all()
+
+    files = [
+        DriveFileResponse(
+            id=f.id,  # type: ignore[arg-type]
+            drive_file_id=f.drive_file_id,
+            name=f.name,
+            mime_type=f.mime_type,
+            size=f.size,
+            modified_time=f.modified_time,
+            last_synced_at=f.last_synced_at,
+            rag_status=f.rag_status,
+            rag_error=f.rag_error,
+        )
+        for f in folder_files
+    ]
+
+    return DriveFolderWithFilesResponse(
+        id=folder.id,  # type: ignore[arg-type]
+        drive_folder_id=folder.drive_folder_id,
+        name=folder.drive_folder_name,
+        parent_id=folder.drive_parent_id,
+        file_count=len(files),
+        last_synced_at=folder.last_synced_at,
+        sync_status=folder.sync_status,
+        files=files,
+    )
+
+
+@router.delete("/folders/{folder_id}")
+def delete_folder(
+    folder_id: int,
+    user_id: int = Depends(require_user_id),
+    session: Session = Depends(get_session),
+) -> dict[str, str]:
+    """Remove a folder and all its files from Sparkth tracking."""
+    folder = session.exec(
+        select(DriveFolder).where(
+            DriveFolder.id == folder_id,
+            DriveFolder.user_id == user_id,
+            DriveFolder.is_deleted == False,  # noqa: E712
+        )
+    ).first()
+    if not folder:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Folder not found")
+
+    folder_files = session.exec(
+        select(DriveFile).where(
+            DriveFile.folder_id == folder.id,
+            DriveFile.is_deleted == False,  # noqa: E712
+        )
+    ).all()
+    for f in folder_files:
+        f.soft_delete()
+        session.add(f)
+
+    folder.soft_delete()
+    session.add(folder)
+    session.commit()
+
+    return {"detail": "Folder removed from Sparkth successfully"}
+
+
+@router.post("/folders/{folder_id}/refresh", response_model=SyncStatusResponse)
+async def refresh_folder(
+    folder_id: int,
+    background_tasks: BackgroundTasks,
+    user_id: int = Depends(require_user_id),
+    session: Session = Depends(get_session),
+) -> SyncStatusResponse:
+    """Refresh folder contents from Google Drive."""
+    folder = session.exec(
+        select(DriveFolder).where(
+            DriveFolder.id == folder_id,
+            DriveFolder.user_id == user_id,
+            DriveFolder.is_deleted == False,  # noqa: E712
+        )
+    ).first()
+    if not folder:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Folder not found")
+
+    client_id, client_secret, _ = get_drive_credentials()
+    access_token = await get_valid_access_token(session, user_id, client_id, client_secret)
+
+    try:
+        await _sync_folder_files(session, folder, user_id, access_token)
+        background_tasks.add_task(process_folder_rag, folder.id, user_id, access_token)  # type: ignore[arg-type]
+
+        return SyncStatusResponse(
+            folder_id=folder.id,  # type: ignore[arg-type]
+            sync_status=folder.sync_status,
+            last_synced_at=folder.last_synced_at,
+        )
+
+    except HTTPException:
+        raise
+    except (ConnectionError, TimeoutError, RuntimeError, ValueError, OSError) as e:
+        folder.sync_status = "error"
+        folder.sync_error = str(e)
+        folder.update_timestamp()
+        session.add(folder)
+        session.commit()
+        session.refresh(folder)
+
+        return SyncStatusResponse(
+            folder_id=folder.id,  # type: ignore[arg-type]
+            sync_status="error",
+            last_synced_at=folder.last_synced_at,
+            error=str(e),
+        )
